+++ uniplot/axis_labels/datetime_label_set.py
@@ -21,8 +21,17 @@ class DatetimeLabelSet(LabelSet):
         log: bool = False,
         vertical_direction: bool = False,
     ):
+        # Note: SI unit prefixes (`unit_as_si`) do not apply to timestamps, so
+        # this class intentionally does not expose that option and lets the base
+        # class default it to `False`.
         super().__init__(
-            labels, x_min, x_max, available_space, unit, log, vertical_direction
+            labels,
+            x_min,
+            x_max,
+            available_space,
+            unit=unit,
+            log=log,
+            vertical_direction=vertical_direction,
         )
         # For convenience, make sure we have the bounds also available as
         # datetime objects.
@@ -33,15 +42,17 @@ class DatetimeLabelSet(LabelSet):
     # private #
     ###########
 
-    def _find_shortest_string_representation(self) -> List[str]:
+    def _find_shortest_string_representation(self, labels=None) -> List[str]:
         """
         This method will find the shortest strings for datetime labels that
         give enough information.
         """
+        if labels is None:
+            labels = self.labels
         # By default, use NumPy's redering functionality
-        short_labels = [str(x) for x in np.datetime_as_string(self.labels, unit="auto")]
+        short_labels = [str(x) for x in np.datetime_as_string(labels, unit="auto")]
         # Remove the date if it is the same in all labels
-        if len(set(np.datetime_as_string(self.labels, unit="D"))) == 1:
+        if len(set(np.datetime_as_string(labels, unit="D"))) == 1:
             short_labels = [t[11:] for t in short_labels]
         # Remove day, and then month, if that is redundant (i.e. all labels
         # have a "-01" at the end)