+++ uniplot/axis_labels/extended_talbot_labels.py
@@ -20,6 +20,7 @@ def extended_talbot_labels(
     available_space: int,
     vertical_direction: bool = False,
     unit: str = "",
+    unit_as_si: bool = False,
     log: bool = False,
     verbose: bool = False,
 ) -> Optional[LabelSet]:
@@ -78,6 +79,7 @@ def extended_talbot_labels(
                     x_max=x_max,
                     available_space=available_space,
                     unit=unit,
+                    unit_as_si=unit_as_si,
                     log=log,
                     vertical_direction=vertical_direction,
                 )